device: extract domain sysfs helper function
Extract the domain related sysfs functions to bolt-sysfs.c
